requirements_consumer.txt
pika
celery==4.4.7
python-decouple